It now works
I changed it so it adds 100px to posX every time they make a new Pessoa
not technically correct but technically correct

+++ Ex_31/Pessoa.pde
@@ -3,7 +3,7 @@ float posX, posY, pWidth, pHeight, btam;
 color cor;
 float mediaWidth, dpWidth,mediaHeight, dpHeight;
 
-  Pessoa(int x) {
+  Pessoa(float x) {
     mediaWidth = 50;    
     dpWidth = 20;
     mediaHeight = 100;
